refactor: controller and service session
