+++ packages/server/src/routes/session/index.ts
@@ -1,6 +1,6 @@
 import { Router } from 'express'
 import loginController from '@/controllers/session/login'
-import RegisterController from '@/controllers/session/register'
+import registerController from '@/controllers/session/register'
 
 const sessionRoutes = Router()
 
@@ -20,7 +20,7 @@ const sessionRoutes = Router()
  *              properties:
  *                message:
  *                  type: string
- *                token: 
+ *                token:
  *                  type: ["string", "null"]
  *              example:
  *                message: Login com sucesso!
@@ -34,7 +34,7 @@ const sessionRoutes = Router()
  *              properties:
  *                message:
  *                  type: string
- *                token: 
+ *                token:
  *                  type: ["string", "null"]
  *            examples:
  *                required-email:
@@ -92,11 +92,39 @@ sessionRoutes.post('/login', loginController)
  *              properties:
  *                message:
  *                  type: string
- *                token: 
+ *                token:
  *                  type: ["string", "null"]
  *              example:
  *                message: Usuário criado com sucesso!
  *                token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpYXQiOjE2NjkyMjkwMjh9.kWzb6evDjKORu9097C5PFrSsfFpL0sxgqlyv-tLDWFc
+ *       409:
+ *         description: Conflict
+ *         content:
+ *          application/json:
+ *            schema:
+ *              type: object
+ *              properties:
+ *                message:
+ *                  type: string
+ *                token:
+ *                  type: ["string", "null"]
+ *              example:
+ *                message: Usuário já existe
+ *                token: null
+ *       404:
+ *         description: Not found
+ *         content:
+ *          application/json:
+ *            schema:
+ *              type: object
+ *              properties:
+ *                message:
+ *                  type: string
+ *                token:
+ *                  type: ["string", "null"]
+ *              example:
+ *                message: Usuário não encontrado!
+ *                token: null
  *       400:
  *         description: Error
  *         content:
@@ -106,18 +134,28 @@ sessionRoutes.post('/login', loginController)
  *              properties:
  *                message:
  *                  type: string
- *                token: 
+ *                token:
  *                  type: ["string", "null"]
  *            examples:
- *                required:
- *                  summary: fields required
+ *                required-name:
+ *                  summary: name required
  *                  value:
- *                    message: Nome, email e senha são obrigatórios
+ *                    message: Nome é obrigatório
  *                    token: null
- *                user:
- *                  summary: user already exists
+ *                required-email:
+ *                  summary: email required
  *                  value:
- *                    message: Usuário já existe
+ *                    message: Email é obrigatório
+ *                    token: null
+ *                email-invalid:
+ *                  summary: email invalid
+ *                  value:
+ *                    message: Email inválido
+ *                    token: null
+ *                required-password:
+ *                  summary: password required
+ *                  value:
+ *                    message: Password é obrigatório
  *                    token: null
  *     consumers:
  *       - application/json
@@ -139,6 +177,6 @@ sessionRoutes.post('/login', loginController)
  *                password:
  *                  type: string
  */
-sessionRoutes.post('/register', RegisterController)
+sessionRoutes.post('/register', registerController)
 
 export { sessionRoutes }
